Create exceptions.py for chipsec exceptions
