+++ chipsec/hal/acpi.py
@@ -33,12 +33,10 @@ from collections import namedtuple
 from chipsec.logger import logger, print_buffer
 from chipsec.file import read_file
 from chipsec.defines import bytestostring
+from chipsec.exceptions import UnimplementedNativeAPIError
 
 from chipsec.hal import acpi_tables, hal_base, uefi
 from chipsec.helper import oshelper
-
-class AcpiRuntimeError (RuntimeError):
-    pass
 
 # ACPI Table Header Format
 ACPI_TABLE_HEADER_FORMAT = '=4sIBB6s8sI4sI'
@@ -400,7 +398,7 @@ class ACPI(hal_base.HALBase):
 
             self.get_table_list_from_SDT(sdt, is_xsdt)
             self.get_DSDT_from_FADT()
-        except oshelper.UnimplementedNativeAPIError:
+        except UnimplementedNativeAPIError:
             # 2. If didn't work, try using get_ACPI_table if a helper implemented
             #    reading ACPI tables via native API which some OS may provide
             if logger().HAL: logger().log( "[acpi] trying to enumerate ACPI tables using get_ACPI_table..." )
@@ -484,7 +482,7 @@ class ACPI(hal_base.HALBase):
                     t_size = self.cs.mem.read_physical_mem_dword( table_address + 4 )
                     t_data = self.cs.mem.read_physical_mem( table_address, t_size )
                     acpi_tables_data.append( t_data )
-            except oshelper.UnimplementedNativeAPIError:
+            except UnimplementedNativeAPIError:
                 # 2. If didn't work, try using get_ACPI_table if a helper implemented
                 #    reading ACPI tables via native API which some OS may provide
                 if logger().HAL: logger().log( "[acpi] trying to extract ACPI table using get_ACPI_table..." )
