+++ chipsec/hal/cpu.py
@@ -35,9 +35,6 @@ VMM_VMWARE  = 0x3
 VMM_KVM     = 0x4
 
 
-class CPURuntimeError (RuntimeError):
-    pass
-
 ########################################################################################################
 #
 # CORES HAL Component
