+++ chipsec/hal/io.py
@@ -34,8 +34,6 @@ usage:
 
 from chipsec.logger import logger
 
-class PortIORuntimeError (RuntimeError):
-    pass
 
 class PortIO:
