+++ chipsec/hal/iommu.py
@@ -27,6 +27,7 @@ Access to IOMMU engines
 """
 
 from chipsec.hal import hal_base, mmio, paging
+from chipec.error import IOMMUError
 
 IOMMU_ENGINE_DEFAULT = 'VTD'
 IOMMU_ENGINE_GFX     = 'GFXVTD'
@@ -37,9 +38,6 @@ IOMMU_ENGINES = {
   IOMMU_ENGINE_DEFAULT: 'VTBAR'
 }
 
-
-class IOMMUError (RuntimeError):
-    pass
 
 class IOMMU(hal_base.HALBase):
