+++ chipsec/helper/linux/linuxhelper.py
@@ -38,7 +38,7 @@ import sys
 import shutil
 
 from chipsec import defines
-from chipsec.helper.oshelper import OsHelperError, HWAccessViolationError, UnimplementedAPIError, UnimplementedNativeAPIError, get_tools_path
+from chipsec.exceptions import OsHelperError, UnimplementedAPIError, UnimplementedNativeAPIError, get_tools_path
 from chipsec.helper.basehelper import Helper
 from chipsec.logger import logger
 import chipsec.file
