+++ chipsec/helper/osx/osxhelper.py
@@ -26,7 +26,7 @@ import shutil
 
 import chipsec
 import chipsec.defines
-from chipsec.helper.oshelper import OsHelperError, HWAccessViolationError, UnimplementedAPIError, UnimplementedNativeAPIError
+from chipsec.exceptions import OsHelperError, UnimplementedAPIError
 from chipsec.helper.basehelper import Helper
 from chipsec.logger import logger
