+++ chipsec/utilcmd/chipset_cmd.py
@@ -27,7 +27,7 @@ usage as a standalone utility:
 """
 
 from chipsec.command    import BaseCommand
-from chipsec.chipset    import UnknownChipsetError
+from chipsec.exceptions    import UnknownChipsetError
 
 # ###################################################################
 #
