+++ chipsec/utilcmd/tpm_cmd.py
@@ -23,6 +23,7 @@
 from chipsec.command import BaseCommand
 from chipsec.hal     import tpm_eventlog
 from chipsec.hal     import tpm
+from chipsec.exceptions   import TpmRuntimeError
 from argparse        import ArgumentParser
 
 class TPMCommand(BaseCommand):
@@ -88,7 +89,7 @@ class TPMCommand(BaseCommand):
         if self.func != self.tpm_parse:
             try:
                 self._tpm = tpm.TPM(self.cs)
-            except tpm.TpmRuntimeError as msg:
+            except TpmRuntimeError as msg:
                 self.logger.log(msg)
                 return
